+++ lib/providers/providers.dart
@@ -2,6 +2,10 @@ import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:letscheck/services/connectivity_service.dart';
 import 'package:package_info_plus/package_info_plus.dart';
+import 'package:rxdart/rxdart.dart';
+import 'package:talker_dio_logger/talker_dio_logger.dart';
+import 'package:talker_flutter/talker_flutter.dart';
+import 'package:dio/dio.dart';
 
 import 'package:checkmk_api/checkmk_api.dart' as cmk_api;
 
@@ -19,11 +23,39 @@ import 'package:letscheck/providers/settings/settings_provider.dart';
 
 export 'package:letscheck/providers/settings/settings_provider.dart';
 
+final talkerProvider = Provider<Talker>((ref) {
+  throw UnimplementedError('Initialize this provider in your app');
+});
+
+final talkerDioLoggerProvider = Provider<TalkerDioLogger>((ref) {
+  return TalkerDioLogger(
+    talker: ref.read(talkerProvider),
+    settings: const TalkerDioLoggerSettings(
+      printRequestData: false,
+      printRequestHeaders: false,
+      printResponseHeaders: false,
+      printResponseMessage: false,
+      printErrorHeaders: false,
+      printResponseData: false,
+      printResponseRedirects: false,
+    ),
+  );
+});
+
 final clientProvider = Provider.family<cmk_api.Client, String>((ref, alias) {
   final settings = ref.watch(settingsProvider
       .select((s) => s.connections.where((c) => c.alias == alias).single));
 
+  final talkerDioLogger = ref.read(talkerDioLoggerProvider);
+
   final client = cmk_api.Client(
+    () {
+      final dio = Dio();
+      dio.interceptors.add(
+        talkerDioLogger,
+      );
+      return dio;
+    },
     cmk_api.ClientSettings(
       baseUrl: settings.baseUrl,
       site: settings.site,
@@ -36,45 +68,41 @@ final clientProvider = Provider.family<cmk_api.Client, String>((ref, alias) {
   return client;
 });
 
-final clientStatesProvider =
-    StreamProvider.family<cmk_api.ConnectionState, String>((ref, alias) {
-  final client = ref.watch(clientProvider(alias));
-  return client.connectionStateStream;
-});
-
 final clientStateProvider =
-    StreamProvider.family<cmk_api.ConnectionState, String>((ref, alias) async* {
+    StreamProvider.family<cmk_api.ConnectionState, String>((ref, alias) {
   final settings = ref.watch(settingsProvider
       .select((s) => s.connections.where((c) => c.alias == alias).single));
 
-  ref.watch(clientStatesProvider(alias));
+  var client = ref.watch(clientProvider(alias));
 
-  final client = ref.watch(clientProvider(alias));
-  final connectivity = await ref.watch(connectivityProvider.future);
-  final currentState = client.currentState;
+  final connectionStateStream = client.connectionStateStream;
 
-  if (settings.wifiOnly) {
-    // Disconnect the client when we are on mobile and wifiOnly is true
-    if (currentState == cmk_api.ConnectionState.connected &&
-        (connectivity.contains(ConnectivityResult.mobile) ||
-            connectivity.contains(ConnectivityResult.none))) {
-      client.disconnect(reason: cmk_api.BaseException(message: "Not on WiFi"));
-    } else if (currentState == cmk_api.ConnectionState.disconnected || currentState == cmk_api.ConnectionState.error) {
-      try {
-        await client.connect();
-      } catch (_) {
-        // Ignore.
+  
+
+  return Rx.combineLatest2(
+      connectionStateStream, ConnectivityService.onConnectivityChanged,
+      (state, connectivity) {
+    final isOnMobileOrNoConnection = connectivity.contains(ConnectivityResult.mobile) ||
+        connectivity.contains(ConnectivityResult.none);
+
+    final paused = client.requestedConnectionState == cmk_api.ConnectionState.paused;
+
+    if (settings.wifiOnly) {
+      if (isOnMobileOrNoConnection && !paused) {
+        // Disconnect the client when we are on mobile and wifiOnly is true
+        client.pause(reason: 'Paused - not on WiFi');
+      } else if (!isOnMobileOrNoConnection && paused) {
+        // Connect the client when we are not on mobile and wifiOnly is true
+        client.connect();
       }
-    } 
-  } else if (currentState == cmk_api.ConnectionState.disconnected || currentState == cmk_api.ConnectionState.error) {
-      try {
-        await client.connect();
-      } catch (_) {
-        // Ignore.
-      }
-  }
+    } else if (state == cmk_api.ConnectionState.initial &&
+        client.requestedConnectionState == cmk_api.ConnectionState.connected) {
+      // Connect the client when the state is initial and we are not on wifi
+      client.connect();
+    }
 
-  yield currentState;
+    return client.connectionState;
+  });
 });
 
 final connectionDataProvider = StateNotifierProvider.family<
@@ -101,7 +129,6 @@ final packageInfoProvider = FutureProvider<PackageInfo>((ref) async {
   return await PackageInfo.fromPlatform();
 });
 
-final javascriptRuntimeProvider =
-    FutureProvider<JavascriptRuntimeWrapper>((ref) async {
-  return await initJavascriptRuntime();
+final javascriptRuntimeProvider = Provider<JavascriptRuntimeWrapper>((ref) {
+  throw UnimplementedError('Initialize the provider in the app');
 });