+++ lib/screen/slim/not_found_screen.dart
@@ -1,15 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'base_slim_screen.dart';
 
 class NotFoundScreen extends ConsumerStatefulWidget {
-  static final route = GoRoute(
-    path: '/error/404',
-    builder: (context, state) => NotFoundScreen(),
-  );
-
   @override
   NotFoundScreenState createState() => NotFoundScreenState();
 }