+++ linux/flutter/generated_plugins.cmake
@@ -7,6 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   flutter_timezone
   screen_retriever_linux
   tray_manager
+  url_launcher_linux
   window_manager
 )