+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - rework connection handling
 - rework connection handling, pt2
+- intodruce talker, rework connection handling, pt3
 
 ## 0.2.0+1272 - 2025-02-08
 ### Added